Move tests from root test_views to appropriate app directory
This splits the tests/test_views.py file so that:

* Tests for old catalogue URLs are in tests/catalogue/, which can
  be removed alongside the app.
* Tests for download pages are moved to a new file in tests/datasets/.
  Since they don't match the pytest approach they're kept in a separate
  file for now.
* Tests for support page are moved to tests/core/, to match the view
  location.
