+++ dataworkspace/dataworkspace/tests/datasets/test_views.py
@@ -108,33 +108,6 @@ def test_request_access_form(client, mocker):
     )
 
 
-def test_old_dataset_url_redirects_to_new_url(client):
-    ds = factories.DataSetFactory.create(published=True)
-    response = client.get(
-        reverse(
-            'catalogue:dataset_fullpath',
-            kwargs={'group_slug': ds.grouping.slug, 'set_slug': ds.slug},
-        )
-    )
-    assert response.status_code == 302
-    assert response['Location'] == ds.get_absolute_url()
-
-
-def test_old_reference_dataset_url_redirects_to_new_url(client):
-    group = factories.DataGroupingFactory.create()
-    rds = factories.ReferenceDatasetFactory.create(
-        group=group, table_name='test_detail_view'
-    )
-    response = client.get(
-        reverse(
-            'catalogue:reference_dataset',
-            kwargs={'group_slug': group.slug, 'reference_slug': rds.slug},
-        )
-    )
-    assert response.status_code == 302
-    assert response['Location'] == rds.get_absolute_url()
-
-
 def test_find_datasets_combines_results(client):
     factories.DataSetFactory.create(published=False, name='Unpublished search dataset')
     ds = factories.DataSetFactory.create(published=True, name='A search dataset')
